Fix: move checkboxLabel
Also made propType for chidren less specific
